LayoutTests/http/wpt/credential-management/credentialscontainer-preventSilentAccess-basics.https.html

<!DOCTYPE html>
<title>Credential Management API: preventSilentAccess() basics.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
promise_test(function() {
    return navigator.credentials.preventSilentAccess()
            .then((result) => {
                assert_equals(result, undefined);
            });
}, "navigator.credentials.preventSilentAccess().");
</script>
